starting tsx and css code for announcement student dashboard

+++ back-end/src/main/java/com/revature/backend/services/AnnouncementsService.java
@@ -1,14 +1,17 @@
 package com.revature.backend.services;
 
 
+import com.revature.backend.exceptions.NoSuchAnnouncementFoundException;
+import com.revature.backend.exceptions.NoSuchCourseException;
 import com.revature.backend.models.Announcements;
 import com.revature.backend.models.Courses;
-import com.revature.backend.models.Users;
 import com.revature.backend.repos.AnnouncementsDAO;
 import com.revature.backend.repos.CoursesDAO;
-import com.revature.backend.repos.UsersDAO;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+
+import java.util.List;
+import java.util.Optional;
 
 @Service
 public class AnnouncementsService {
@@ -21,16 +24,31 @@ public class AnnouncementsService {
         this.coursesDAO = coursesDAO;
     }
 
-    public Announcements createAnnouncement(Integer courseId, Announcements announcement) {
+    public Announcements createAnnouncement(Integer courseId, Announcements announcement) throws NoSuchCourseException{
         Courses course = coursesDAO.findById(courseId)
-                .orElseThrow(() -> new IllegalArgumentException("Course not found"));
+                .orElseThrow(() -> new NoSuchCourseException("No course found with ID: " + courseId));
 
         announcement.setCourse(course);
 
         return announcementsDAO.save(announcement);
     }
 
-    public void deleteAnnouncement(Integer announcementId) {
+    public Announcements deleteAnnouncement(Integer announcementId) throws NoSuchAnnouncementFoundException {
+        Optional<Announcements> announcement = announcementsDAO.findById(announcementId);
+        if(announcement.isEmpty()){
+            throw new NoSuchAnnouncementFoundException("No announcement found with ID: " + announcementId);
+        }
         announcementsDAO.deleteById(announcementId);
+        return announcement.get();
+    }
+
+    public List<Announcements> getAllAnnouncements(Integer courseId) {
+        return announcementsDAO.findByCourse_CourseId(courseId);
+    }
+
+    public Announcements getAnnouncementById(Integer announcementId) throws NoSuchAnnouncementFoundException {
+        return announcementsDAO.findById(announcementId)
+                .orElseThrow(() -> new NoSuchAnnouncementFoundException("No announcement found with ID: " + announcementId));
+
     }
 }
